Rename `datatransfer` to `io` and refactor io-related tests
