+++ pixeltable/globals.py
@@ -394,68 +394,3 @@ def get_path(schema_obj: catalog.SchemaObject) -> str:
         dir_id = dir._dir_id
     path_elements.append(schema_obj._name)
     return '.'.join(path_elements)
-
-
-def import_huggingface_dataset(
-    table_path: str,
-    dataset: Union['datasets.Dataset', 'datasets.DatasetDict'],
-    *,
-    column_name_for_split: Optional[str] = 'split',
-    schema_override: Optional[dict[str, Any]] = None,
-    **kwargs
-) -> catalog.InsertableTable:
-    """Create a new `InsertableTable` from a Huggingface dataset, or dataset dict with multiple splits.
-        Requires datasets library to be installed.
-
-    Args:
-        path_str: Path to the table.
-        dataset: Huggingface datasts.Dataset or datasts.DatasetDict to insert into the table.
-        column_name_for_split: column name to use for split information. If None, no split information will be stored.
-        schema_override: Optional dictionary mapping column names to column type to override the corresponding defaults from
-        `pixeltable.utils.hf_datasets.huggingface_schema_to_pixeltable_schema`. The column type should be a pixeltable ColumnType.
-        For example, {'col_vid': VideoType()}, rather than {'col_vid': StringType()}.
-
-        kwargs: Additional arguments to pass to `create_table`.
-
-    Returns:
-        The newly created table. The table will have loaded the data from the dataset.
-    """
-    from pixeltable.utils import hf_datasets
-
-    return hf_datasets.import_huggingface_dataset(
-        table_path,
-        dataset,
-        column_name_for_split=column_name_for_split,
-        schema_override=schema_override,
-        **kwargs,
-    )
-
-
-def import_parquet(
-    table_path: str,
-    *,
-    parquet_path: str,
-    schema_override: Optional[dict[str, Any]] = None,
-    **kwargs,
-) -> catalog.InsertableTable:
-    """Create a new `InsertableTable` from a Parquet file or set of files. Requires pyarrow to be installed.
-    Args:
-        path_str: Path to the table within pixeltable.
-        parquet_path: Path to an individual Parquet file or directory of Parquet files.
-        schema_override: Optional dictionary mapping column names to column type to override the default
-                        schema inferred from the Parquet file. The column type should be a pixeltable ColumnType.
-                        For example, {'col_vid': VideoType()}, rather than {'col_vid': StringType()}.
-                        Any fields not provided explicitly will map to types with `pixeltable.utils.parquet.parquet_schema_to_pixeltable_schema`
-        kwargs: Additional arguments to pass to `create_table`.
-
-    Returns:
-        The newly created table. The table will have loaded the data from the Parquet file(s).
-    """
-    from pixeltable.utils import parquet
-
-    return parquet.import_parquet(
-        table_path=table_path,
-        parquet_path=parquet_path,
-        schema_override=schema_override,
-        **kwargs,
-    )
