+++ pixeltable/io/__init__.py
@@ -1,1 +1,3 @@
+from .hf_datasets import import_huggingface_dataset
 from .pandas import import_csv, import_excel, import_pandas
+from .parquet import import_parquet
