+++ pixeltable/tests/io/test_pandas.py
@@ -8,11 +8,10 @@ from ..utils import skip_test_if_not_installed
 
 class TestPandas:
 
-    def test_pandas_csv(self, test_client: pxt.Client) -> None:
-        cl = test_client
-        from pixeltable.io.pandas import import_csv
+    def test_pandas_csv(self, reset_db) -> None:
+        from pixeltable.io import import_csv
 
-        t1 = import_csv(cl, 'online_foods', 'pixeltable/tests/data/datasets/onlinefoods.csv')
+        t1 = import_csv('online_foods', 'pixeltable/tests/data/datasets/onlinefoods.csv')
         assert t1.count() == 388
         assert t1.column_types() == {
             'Age': pxt.IntType(),
@@ -31,7 +30,7 @@ class TestPandas:
         }
         assert t1.select(t1.Age).limit(5).collect()['Age'][:5] == [20, 24, 22, 22, 22]
 
-        t2 = import_csv(cl, 'ibm', 'pixeltable/tests/data/datasets/classeurIBM.csv')
+        t2 = import_csv('ibm', 'pixeltable/tests/data/datasets/classeurIBM.csv')
         assert t2.count() == 4263
         assert t2.column_types() == {
             'Date': pxt.StringType(),
@@ -44,7 +43,6 @@ class TestPandas:
         }
 
         t3 = import_csv(
-            cl,
             'edge_cases',
             'pixeltable/tests/data/datasets/edge-cases.csv',
             parse_dates=['ts', 'ts_n']
@@ -70,14 +68,12 @@ class TestPandas:
         assert result_set['ts'] == [datetime.datetime(2024, 5, n) for n in range(3, 7)]
         assert result_set['ts_n'] == [datetime.datetime(2024, 5, 3), None, None, datetime.datetime(2024, 5, 6)]
 
-    def test_pandas_images(self, test_client: pxt.Client) -> None:
+    def test_pandas_images(self, reset_db) -> None:
         skip_test_if_not_installed('boto3')  # This test relies on s3 URLs
-        cl = test_client
         from pixeltable.io.pandas import import_csv
 
         # Test overriding string type to images
         t4 = import_csv(
-            cl,
             'images',
             'pixeltable/tests/data/datasets/images.csv',
             schema={'name': pxt.StringType(), 'image': pxt.ImageType(nullable=True)}
@@ -90,18 +86,17 @@ class TestPandas:
         result_set = t4.select(t4.image.width).collect()
         assert result_set['width'] == [1024, 962, 1024, None]
 
-    def test_pandas_excel(self, test_client: pxt.Client) -> None:
+    def test_pandas_excel(self, reset_db) -> None:
         skip_test_if_not_installed('openpyxl')
         from pixeltable.io.pandas import import_excel
-        cl = test_client
 
-        t4 = import_excel(cl, 'fin_sample', 'pixeltable/tests/data/datasets/Financial Sample.xlsx')
+        t4 = import_excel('fin_sample', 'pixeltable/tests/data/datasets/Financial Sample.xlsx')
         assert t4.count() == 700
         assert t4.column_types()['Date'] == pxt.TimestampType()
         entry = t4.df().limit(1).collect()[0]
         assert entry['Date'] == datetime.datetime(2014, 1, 1, 0, 0)
 
-        t5 = import_excel(cl, 'sale_data', 'pixeltable/tests/data/datasets/SaleData.xlsx')
+        t5 = import_excel('sale_data', 'pixeltable/tests/data/datasets/SaleData.xlsx')
         assert t5.count() == 45
         assert t5.column_types()['OrderDate'] == pxt.TimestampType(nullable=True)
         # Ensure valid mapping of 'NaT' -> None
