+++ pixeltable/tests/test_table.py
@@ -137,100 +137,6 @@ class TestTable:
         tbl.revert()
         assert tbl.num_retained_versions == num_retained_versions
 
-    def test_import_parquet(self, reset_db, tmp_path: pathlib.Path) -> None:
-        skip_test_if_not_installed('pyarrow')
-        import pyarrow as pa
-        from pixeltable.utils.arrow import iter_tuples
-
-        parquet_dir = tmp_path / 'test_data'
-        parquet_dir.mkdir()
-        make_test_arrow_table(parquet_dir)
-
-        tab = pxt.import_parquet('test_parquet', parquet_path=str(parquet_dir))
-        assert 'test_parquet' in pxt.list_tables()
-        assert tab is not None
-        num_elts = tab.count()
-        arrow_tab: pa.Table = pa.parquet.read_table(str(parquet_dir))
-        assert num_elts == arrow_tab.num_rows
-        assert set(tab.column_names()) == set(arrow_tab.column_names)
-
-        result_set = tab.order_by(tab.c_id).collect()
-        column_types = tab.column_types()
-
-        for tup, arrow_tup in zip(result_set, iter_tuples(arrow_tab)):
-            assert tup['c_id'] == arrow_tup['c_id']
-            for col, val in tup.items():
-                if val is None:
-                    assert arrow_tup[col] is None
-                    continue
-
-                if column_types[col].is_array_type():
-                    assert (val == arrow_tup[col]).all()
-                else:
-                    assert val == arrow_tup[col]
-
-    def test_import_huggingface_dataset(self, reset_db, tmp_path: pathlib.Path) -> None:
-        skip_test_if_not_installed('datasets')
-        import datasets
-
-        test_cases = [
-            # { # includes a timestamp. 20MB for specific slice
-            # Disbled this test case because download is failing, and its not critical.
-            #     'dataset_name': 'c4',
-            #     # see https://huggingface.co/datasets/allenai/c4/blob/main/realnewslike/c4-train.00000-of-00512.json.gz
-            #     'dataset': datasets.load_dataset(
-            #         "allenai/c4",
-            #         data_dir="realnewslike",
-            #         data_files="c4-train.00000-of-00512.json.gz",
-            #         split='train[:1000]',
-            #         cache_dir=tmp_path
-            #     ),
-            # },
-            {  # includes an embedding (array type), common in a few RAG datasets.
-                'dataset_name': 'cohere_wikipedia',
-                'dataset': datasets.load_dataset("Cohere/wikipedia-2023-11-embed-multilingual-v3",
-                                                 data_dir='cr').select_columns(['url', 'title', 'text', 'emb']),
-                # column with name `_id`` is not currently allowed by pixeltable rules,
-                # so filter out that column.
-                # cr subdir has a small number of rows, avoid running out of space in CI runner
-                # see https://huggingface.co/datasets/Cohere/wikipedia-2023-11-embed-multilingual-v3/tree/main/cr
-                'schema_override': {'emb': ArrayType((1024,), dtype=FloatType(), nullable=False)}
-            },
-            # example of dataset dictionary with multiple splits
-            {
-                'dataset_name': 'rotten_tomatoes',
-                'dataset': datasets.load_dataset("rotten_tomatoes"),
-            },
-        ]
-
-        # test a column name for splits other than the default of 'split'
-        split_column_name = 'my_split_col'
-        for rec in test_cases:
-            dataset_name = rec['dataset_name']
-            hf_dataset = rec['dataset']
-
-            tab = pxt.import_huggingface_dataset(
-                dataset_name,
-                hf_dataset,
-                column_name_for_split=split_column_name,
-                schema_override=rec.get('schema_override', None),
-            )
-            if isinstance(hf_dataset, datasets.Dataset):
-                assert_hf_dataset_equal(hf_dataset, tab.df(), split_column_name)
-            elif isinstance(hf_dataset, datasets.DatasetDict):
-                assert tab.count() == sum(hf_dataset.num_rows.values())
-                assert split_column_name in tab.column_names()
-
-                for dataset_name in hf_dataset:
-                    df = tab.where(tab.my_split_col == dataset_name)
-                    assert_hf_dataset_equal(hf_dataset[dataset_name], df, split_column_name)
-            else:
-                assert False
-
-        with pytest.raises(excs.Error) as exc_info:
-            pxt.import_huggingface_dataset('test', {})
-        assert 'type(dataset)' in str(exc_info.value)
-
     def test_image_table(self, reset_db) -> None:
         n_sample_rows = 20
         schema = {
